Modernize and fix memory unsafe and compiler warnings

+++ src/RPGTypes.cpp
@@ -50,12 +50,12 @@ StringTemp operator + (const FixedTemp &fStr1, const FixedTemp &fStr2)
 }
 StringTemp operator + (const char *str, const FixedTemp &fStr2)
 {
-	FixedTemp fStr1((char *)str, strlen(str));
+	FixedTemp fStr1((char *)str, static_cast<int>(strlen(str)));
 	return fStr1 + fStr2;
 }
 StringTemp operator + (const FixedTemp &fStr1, const char *str)
 {
-	FixedTemp fStr2((char *)str, strlen(str));
+	FixedTemp fStr2((char *)str, static_cast<int>(strlen(str)));
 	return fStr1 + fStr2;
 }
